Separate dev into edge and gcs modules
