+++ src/commands/dev/edge/server.rs
@@ -1,0 +1,3 @@
+pub(super) async fn serve() -> Result<(), failure::Error> {
+  unimplemented!()
+}
