+++ src/commands/dev/edge/setup.rs
@@ -1,0 +1,3 @@
+pub(super) fn get_preview_id() -> Result<String, failure::Error> {
+    unimplemented!()
+}
