+++ src/commands/dev/gcs/watch.rs
@@ -1,8 +1,9 @@
 use std::sync::{mpsc, Arc, Mutex};
 
 use crate::commands;
-use crate::commands::dev::get_preview_id;
+use crate::commands::dev::gcs::setup::get_preview_id;
 use crate::commands::dev::server_config::ServerConfig;
+use crate::commands::watch_and_build;
 
 use crate::settings::global_user::GlobalUser;
 use crate::settings::toml::Target;
@@ -16,7 +17,7 @@ pub fn watch_for_changes(
     verbose: bool,
 ) -> Result<(), failure::Error> {
     let (sender, receiver) = mpsc::channel();
-    commands::watch_and_build(&target, Some(sender))?;
+    watch_and_build(&target, Some(sender))?;
 
     while let Ok(_) = receiver.recv() {
         let user = user.clone();
